+++ src/ansys/fluent/core/launcher/launch_options.py
@@ -139,7 +139,8 @@ class FluentMode(FluentEnum):
     def _default(self):
         return self.SOLVER
 
-    def _get_enum_map(self):
+    @classmethod
+    def _get_enum_map(cls):
         from ansys.fluent.core.session.meshing import Meshing
         from ansys.fluent.core.session.pure_meshing import PureMeshing
         from ansys.fluent.core.session.solver import Solver
@@ -147,13 +148,18 @@ class FluentMode(FluentEnum):
         from ansys.fluent.core.session.solver_icing import SolverIcing
 
         return {
-            self.MESHING: Meshing,
-            self.PURE_MESHING: PureMeshing,
-            self.SOLVER: Solver,
-            self.SOLVER_ICING: SolverIcing,
-            self.SOLVER_AERO: SolverAero,
-            self.PRE_POST: Solver,
+            cls.MESHING: Meshing,
+            cls.PURE_MESHING: PureMeshing,
+            cls.SOLVER: Solver,
+            cls.SOLVER_ICING: SolverIcing,
+            cls.SOLVER_AERO: SolverAero,
+            cls.PRE_POST: Solver,
         }
+
+    @classmethod
+    def from_session_class(cls, session_class):
+        """Get the FluentMode corresponding to a session class."""
+        return {v: k for k, v in cls._get_enum_map().items()}.get(session_class)
 
     @staticmethod
     def is_meshing(mode: "FluentMode") -> bool:
