+++ src/ansys/fluent/core/session/session.py
@@ -33,14 +33,26 @@ from functools import cached_property
 import json
 import logging
 import os
-from typing import Any
+from typing import TYPE_CHECKING, Any, Literal, overload
 import warnings
 import weakref
 
 from deprecated.sphinx import deprecated
+from typing_extensions import Unpack
 
 from ansys.fluent.core._types import PathType
 from ansys.fluent.core.fields.live_field_data import LiveFieldData, ZoneInfo, _FieldInfo
+from ansys.fluent.core.launcher.launch_options import FluentMode
+
+if TYPE_CHECKING:
+    from ansys.fluent.core.launcher.standalone_launcher import (
+        StandaloneArgsWithoutDryRunMode,
+    )
+    from ansys.fluent.core.launcher.container_launcher import (
+        ContainerArgsWithoutDryRunMode,
+    )
+    from ansys.fluent.core.launcher.pim_launcher import PIMArgsWithoutMode
+
 from ansys.fluent.core.fluent_connection import FluentConnection
 from ansys.fluent.core.journaling import Journal
 from ansys.fluent.core.pyfluent_warnings import (
@@ -99,6 +111,31 @@ class BaseSession:
     exit()
         Close the Fluent connection and exit Fluent.
     """
+
+    @classmethod
+    def _validate_mode_not_in_kwargs(
+        cls, kwargs: dict[str, Any], method_name: str
+    ) -> None:
+        """Validate that 'mode' is not in kwargs.
+
+        Parameters
+        ----------
+        kwargs : dict[str, Any]
+            Keyword arguments to validate
+        method_name : str
+            Name of calling method (e.g., 'from_install')
+
+        Raises
+        ------
+        ValueError
+            If 'mode' is present in kwargs
+        """
+        if "mode" in kwargs:
+            raise ValueError(
+                f"Cannot specify 'mode' in {cls.__name__}.{method_name}(). "
+                "The mode is determined by the session class. "
+                f"You are already using {cls.__name__}, which sets mode to {FluentMode.from_session_class(cls)}."
+            )
 
     def __new__(cls, *args, **kwargs):
         if cls is BaseSession:
@@ -491,6 +528,371 @@ class BaseSession:
     @property
     def _is_beta_enabled(self):
         return self.application_runtime.is_beta_enabled()
+
+    @overload
+    @classmethod
+    def from_install(
+        cls,
+        *,
+        dry_run: Literal[False] = False,
+        **kwargs: Unpack["StandaloneArgsWithoutDryRunMode"],
+    ) -> "BaseSession": ...
+
+    @overload
+    @classmethod
+    def from_install(
+        cls,
+        *,
+        dry_run: Literal[True],
+        **kwargs: Unpack["StandaloneArgsWithoutDryRunMode"],
+    ) -> tuple[str, str]: ...
+
+    @classmethod
+    def from_install(  # pylint: disable=missing-param-doc
+        cls,
+        *,
+        dry_run: bool = False,
+        **kwargs: Unpack["StandaloneArgsWithoutDryRunMode"],
+    ) -> "BaseSession | tuple[str, str]":
+        """
+        Launch a Fluent session in standalone mode.
+
+        Parameters
+        ----------
+        ui_mode : UIMode or str, optional
+            Defines the user interface mode for Fluent. Accepts either a ``UIMode`` value
+            or a corresponding string such as ``"no_gui"``, ``"hidden_gui"``, or ``"gui"``.
+        graphics_driver : FluentWindowsGraphicsDriver or FluentLinuxGraphicsDriver
+            Specifies the graphics driver for Fluent. Options are from the ``FluentWindowsGraphicsDriver`` enum
+            (for Windows) or the ``FluentLinuxGraphicsDriver`` enum (for Linux).
+        product_version : FluentVersion or str or float or int, optional
+            Indicates the version of Ansys Fluent to launch. For example, to use version 2025 R1, pass
+            ``FluentVersion.v251``, ``"25.1.0"``, ``"25.1"``, ``25.1``, or ``251``. Defaults to ``None``,
+            which uses the newest installed version.
+        dimension : Dimension or int, optional
+            Specifies the geometric dimensionality of the Fluent simulation. Defaults to ``None``,
+            which corresponds to ``Dimension.THREE``. Acceptable values are from the ``Dimension`` enum
+            (``Dimension.TWO`` or ``Dimension.THREE``) or integers ``2`` and ``3``.
+        precision : Precision or str, optional
+            Defines the floating point precision. Defaults to ``None``, which corresponds to
+            ``Precision.DOUBLE``. Acceptable values are from the ``Precision`` enum (``Precision.SINGLE``
+            or ``Precision.DOUBLE``) or strings ``"single"`` and ``"double"``.
+        processor_count : int, optional
+            Specifies the number of processors to use. Defaults to ``None``, which uses 1 processor.
+            In job scheduler environments, this value limits the total number of allocated cores.
+        journal_file_names : str or list of str, optional
+            Path(s) to a Fluent journal file(s) that Fluent will execute. Defaults to ``None``.
+        start_timeout : int, optional
+            Maximum time in seconds allowed for connecting to the Fluent server. Defaults to 100 seconds.
+        additional_arguments : str, optional
+            Additional command-line arguments for Fluent, formatted as they would be on the command line.
+        env : dict[str, str], optional
+            A mapping for modifying environment variables in Fluent. Defaults to ``None``.
+        cleanup_on_exit : bool, optional
+            Determines whether to shut down the connected Fluent session when exiting PyFluent or calling
+            the session's `exit()` method. Defaults to True.
+        dry_run : bool, optional
+            If True, does not launch Fluent but prints configuration information instead. The `call()` method
+            returns a tuple containing the launch string and server info file name. Defaults to False.
+        start_transcript : bool, optional
+            Indicates whether to start streaming the Fluent transcript in the client. Defaults to True;
+            streaming can be controlled via `transcript.start()` and `transcript.stop()` methods on the session object.
+        case_file_name : :class:`os.PathLike` or str, optional
+            Name of the case file to read into the Fluent session. Defaults to None.
+        case_data_file_name : :class:`os.PathLike` or str, optional
+            Name of the case data file. If both case and data files are provided, they are read into the session.
+        lightweight_mode : bool, optional
+            If True, runs in lightweight mode where mesh settings are read into a background solver session,
+            replacing it once complete. This parameter is only applicable when `case_file_name` is provided; defaults to False.
+        py : bool, optional
+            If True, runs Fluent in Python mode. Defaults to None.
+        gpu : bool, optional
+            If True, starts Fluent with GPU Solver enabled.
+        cwd : :class:`os.PathLike` or str, optional
+            Working directory for the Fluent client.
+        fluent_path: :class:`os.PathLike` or str, optional
+            User-specified path for Fluent installation.
+        topy :  bool or str, optional
+            A flag indicating whether to write equivalent Python journals from provided journal files; can also specify
+            a filename for the new Python journal.
+        start_watchdog : bool, optional
+            When `cleanup_on_exit` is True, defaults to True; an independent watchdog process ensures that any local
+            GUI-less Fluent sessions started by PyFluent are properly closed when the current Python process ends.
+        file_transfer_service : Any
+            Service for uploading/downloading files to/from the server.
+
+        Raises
+        ------
+        UnexpectedKeywordArgument
+            If an unexpected keyword argument is provided.
+
+        ValueError
+            If 'mode' is passed in kwargs along with container parameter.
+
+        Notes
+        -----
+        In job scheduler environments (e.g., SLURM, LSF, PBS), resources and compute nodes are allocated,
+        and core counts are queried from these environments before being passed to Fluent.
+        """
+        cls._validate_mode_not_in_kwargs(kwargs, "from_install")
+        from ansys.fluent.core.launcher.standalone_launcher import StandaloneLauncher
+
+        launcher = StandaloneLauncher(
+            **kwargs, dry_run=dry_run, mode=FluentMode.from_session_class(cls)
+        )
+        return launcher()
+
+    @overload
+    @classmethod
+    def from_container(
+        cls,
+        *,
+        dry_run: Literal[False] = False,
+        **kwargs: Unpack["ContainerArgsWithoutDryRunMode"],
+    ) -> "BaseSession": ...
+
+    @overload
+    @classmethod
+    def from_container(
+        cls,
+        *,
+        dry_run: Literal[True],
+        **kwargs: Unpack["ContainerArgsWithoutDryRunMode"],
+    ) -> dict[str, Any]: ...
+
+    @classmethod
+    def from_container(  # pylint: disable=missing-param-doc
+        cls,
+        *,
+        dry_run: bool = False,
+        **kwargs: Unpack["ContainerArgsWithoutDryRunMode"],
+    ) -> "BaseSession | dict[str, Any]":
+        """
+        Launch a Fluent session in container mode.
+
+        Parameters
+        ----------
+        ui_mode : UIMode or str, optional
+            Defines the user interface mode for Fluent. Accepts either a ``UIMode`` value
+            or a corresponding string such as ``"no_gui"``, ``"hidden_gui"``, or ``"gui"``.
+        graphics_driver : FluentWindowsGraphicsDriver or FluentLinuxGraphicsDriver
+            Specifies the graphics driver for Fluent. Options are from the ``FluentWindowsGraphicsDriver`` enum
+            (for Windows) or the ``FluentLinuxGraphicsDriver`` enum (for Linux).
+        product_version :  FluentVersion or str or float or int, optional
+            Indicates the version of Ansys Fluent to launch. For example, to use version 2025 R1, pass
+            any of ``FluentVersion.v251``, ``"25.1.0"``, ``"25.1"``, ``25.1``, or ``251``. Defaults to ``None``,
+            which uses the newest installed version.
+        dimension : Dimension or int, optional
+            Specifies the geometric dimensionality of the Fluent simulation. Defaults to ``None``,
+            which corresponds to ``Dimension.THREE``. Acceptable values include ``Dimension.TWO``,
+            ``Dimension.THREE``, or integers ``2`` and ``3``.
+        precision : Precision or str, optional
+            Defines the floating point precision. Defaults to ``None``, which corresponds to
+            ``Precision.DOUBLE``. Acceptable values include ``Precision.SINGLE``,
+            ``Precision.DOUBLE``, or strings ``"single"`` and ``"double"``.
+        processor_count : int, optional
+            Specifies the number of processors to use. Defaults to ``None``, which uses 1 processor.
+            In job scheduler environments, this value limits the total number of allocated cores.
+        start_timeout : int, optional
+            Maximum allowable time in seconds for connecting to the Fluent server. Defaults to 100 seconds.
+        additional_arguments : str, optional
+            Additional command-line arguments for Fluent, formatted as they would be on the command line.
+        container_dict : dict, optional
+            Configuration dictionary for launching Fluent inside a Docker container. See also
+            :mod:`~ansys.fluent.core.launcher.fluent_container`.
+        dry_run : bool, optional
+            If True, does not launch Fluent but prints configuration information instead. If dry running a
+            container start, this method will return the configured ``container_dict``. Defaults to False.
+        cleanup_on_exit : bool
+            Determines whether to shut down the connected Fluent session upon exit or when calling
+            the session's `exit()` method. Defaults to True.
+        start_transcript : bool
+            Indicates whether to start streaming the Fluent transcript in the client. Defaults to True;
+            streaming can be controlled via `transcript.start()` and `transcript.stop()` methods on the session object.
+        py : bool, optional
+            If True, runs Fluent in Python mode. Defaults to None.
+        gpu : bool, optional
+            If True, starts Fluent with GPU Solver enabled.
+        start_watchdog : bool, optional
+            If True and `cleanup_on_exit` is True, an independent watchdog process is run to ensure that any local
+            GUI-less Fluent sessions started by PyFluent are properly closed when the current Python process ends.
+        file_transfer_service : Any, optional
+            Service for uploading/downloading files to/from the server.
+        use_docker_compose: bool
+            Whether to use Docker Compose to launch Fluent.
+        use_podman_compose: bool
+            Whether to use Podman Compose to launch Fluent.
+        certificates_folder : str, optional
+            Path to the folder containing TLS certificates for Fluent's gRPC server.
+        insecure_mode : bool, optional
+            If True, Fluent's gRPC server will be started in insecure mode without TLS.
+            This mode is not recommended. For more details on the implications
+            and usage of insecure mode, refer to the Fluent documentation.
+
+        Returns
+        -------
+        Meshing | PureMeshing | Solver | SolverIcing | dict
+            Session object or configuration dictionary if ``dry_run`` is True.
+
+        Raises
+        ------
+        UnexpectedKeywordArgument
+            If an unexpected keyword argument is provided.
+
+        ValueError
+            If 'mode' is passed in kwargs.
+
+        Notes
+        -----
+        In job scheduler environments (e.g., SLURM, LSF, PBS), resources and compute nodes are allocated,
+        and core counts are queried from these environments before being passed to Fluent.
+        """
+        cls._validate_mode_not_in_kwargs(kwargs, "from_container")
+        from ansys.fluent.core.launcher.container_launcher import DockerLauncher
+
+        launcher = DockerLauncher(
+            **kwargs, dry_run=dry_run, mode=FluentMode.from_session_class(cls)
+        )
+        return launcher()
+
+    @classmethod
+    def from_pim(  # pylint: disable=missing-param-doc
+        cls,
+        **kwargs: Unpack["PIMArgsWithoutMode"],
+    ) -> "BaseSession":
+        """
+        Launch a Fluent session in `PIM <https://pypim.docs.pyansys.com/version/stable/>`_ mode.
+
+        Parameters
+        ----------
+        ui_mode : UIMode or str, optional
+            Defines the user interface mode for Fluent. Accepts either a ``UIMode`` value
+            or a corresponding string such as ``"no_gui"``, ``"hidden_gui"``, or ``"gui"``.
+        graphics_driver : FluentWindowsGraphicsDriver or FluentLinuxGraphicsDriver
+            Specifies the graphics driver for Fluent. Options are from the ``FluentWindowsGraphicsDriver`` enum
+            (for Windows) or the ``FluentLinuxGraphicsDriver`` enum (for Linux).
+        product_version : FluentVersion or str or float or int, optional
+            Indicates the version of Ansys Fluent to launch. For example, to use version 2025 R1, pass
+            any of ``FluentVersion.v251``, ``"25.1.0"``, ``"25.1"``, ``25.1``, or ``251``. Defaults to ``None``,
+            which uses the newest installed version.
+        dimension : Dimension or int, optional
+            Specifies the geometric dimensionality of the Fluent simulation. Defaults to ``None``,
+            which corresponds to ``Dimension.THREE``. Acceptable values include ``Dimension.TWO``,
+            ``Dimension.THREE``, or integers ``2`` and ``3``.
+        precision : Precision or str, optional
+            Defines the floating point precision. Defaults to ``None``, which corresponds to
+            ``Precision.DOUBLE``. Acceptable values include ``Precision.SINGLE``,
+            ``Precision.DOUBLE``, or strings ``"single"`` and ``"double"``.
+        processor_count : int, optional
+            Specifies the number of processors to use. Defaults to ``None``, which uses 1 processor.
+            In job scheduler environments, this value limits the total number of allocated cores.
+        start_timeout : int, optional
+            Maximum allowable time in seconds for connecting to the Fluent server. Defaults to 100 seconds.
+        additional_arguments : str, optional
+            Additional command-line arguments for Fluent, formatted as they would be on the command line.
+        cleanup_on_exit : bool
+            Determines whether to shut down the connected Fluent session upon exit or when calling
+            the session's `exit()` method. Defaults to True.
+        dry_run : bool, optional
+            If True, does not launch Fluent but prints configuration information instead. If dry running a
+            PIM start, this method will return a configuration dictionary. Defaults to False.
+        start_transcript : bool
+            Indicates whether to start streaming the Fluent transcript in the client. Defaults to True;
+            streaming can be controlled via `transcript.start()` and `transcript.stop()` methods on the session object.
+        gpu : bool, optional
+            If True, starts Fluent with GPU Solver enabled.
+        start_watchdog : bool, optional
+            If True and `cleanup_on_exit` is True, an independent watchdog process is run to ensure that any local
+            GUI-less Fluent sessions started by PyFluent are properly closed when the current Python process ends.
+        file_transfer_service : Any, optional
+            Service for uploading/downloading files to/from the server.
+
+        Returns
+        -------
+        Union[Meshing, PureMeshing, Solver, SolverIcing, dict]
+            Session object or configuration dictionary if ``dry_run`` is True.
+
+        Raises
+        ------
+        UnexpectedKeywordArgument
+            If an unexpected keyword argument is provided.
+
+        Notes
+        -----
+        In job scheduler environments (e.g., SLURM, LSF, PBS), resources and compute nodes are allocated,
+        and core counts are queried from these environments before being passed to Fluent.
+        """
+        from ansys.fluent.core.launcher.pim_launcher import PIMLauncher
+
+        launcher = PIMLauncher(**kwargs, mode=FluentMode.from_session_class(cls))
+        return launcher()
+
+    @classmethod
+    def from_connection(
+        cls,
+        ip: str | None = None,
+        port: int | None = None,
+        server_info_file_name: str | None = None,
+        password: str | None = None,
+        allow_remote_host: bool = False,
+        certificates_folder: str | None = None,
+        insecure_mode: bool = False,
+    ):
+        """Connect to an existing Fluent server instance.
+
+        Parameters
+        ----------
+        ip : str, optional
+            IP address for connecting to an existing Fluent instance. The
+            IP address defaults to ``"127.0.0.1"``. You can also use the environment
+            variable ``PYFLUENT_FLUENT_IP=<ip>`` to set this parameter.
+            The explicit value of ``ip`` takes precedence over ``PYFLUENT_FLUENT_IP=<ip>``.
+        port : int, optional
+            Port to listen on for an existing Fluent instance. You can use the
+            environment variable ``PYFLUENT_FLUENT_PORT=<port>`` to set a default
+            value. The explicit value of ``port`` takes precedence over
+            ``PYFLUENT_FLUENT_PORT=<port>``.
+        server_info_file_name: str
+            Path to server-info file written out by Fluent server. The default is
+            ``None``. PyFluent uses the connection information in the file to
+            connect to a running Fluent session.
+        password : str, optional
+            Password to connect to existing Fluent instance.
+        allow_remote_host : bool, optional
+            Whether to allow connecting to a remote Fluent instance.
+        certificates_folder : str, optional
+            Path to the folder containing TLS certificates for Fluent's gRPC server.
+        insecure_mode : bool, optional
+            If True, Fluent's gRPC server will be connected in insecure mode without TLS.
+            This mode is not recommended. For more details on the implications
+            and usage of insecure mode, refer to the Fluent documentation.
+
+        Raises
+        ------
+        TypeError
+            If the session type does not match the expected session type.
+        """
+        from ansys.fluent.core.launcher.launcher import connect_to_fluent
+
+        session = connect_to_fluent(
+            ip=ip,
+            port=port,
+            server_info_file_name=server_info_file_name,
+            password=password,
+            allow_remote_host=allow_remote_host,
+            certificates_folder=certificates_folder,
+            insecure_mode=insecure_mode,
+        )
+
+        expected = "Solver" if cls.__name__ == "PrePost" else cls.__name__
+        actual = session.__class__.__name__
+
+        if actual != expected:
+            raise TypeError(
+                f"Session type mismatch: expected {expected}, got {actual}."
+            )
+
+        return session
 
 
 class Fields:
